Add variations of intent cue
Problem: The commands in this file are extremely limited in the range of natural vocabulary and available language structure commonly used in day-to-day speech.

Rational: To be front-runner of the next generation of AI-powered PDA, one will need to be understood when speaking in day-to-day English (or other languages). One does not simply speak in the imperative at all times. Additional command forms ("cues") are required for such use.

Changes will:
* add new grammatical variations of the original cues (to make interaction more fluid);
* add polite variations to all commands with which politeness modifiers ("please") create valid human-understood semantics.